+++ collector-bitmovin-player/src/main/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapter.java
@@ -15,6 +15,7 @@ import com.bitmovin.analytics.enums.VideoStartFailedReason;
 import com.bitmovin.analytics.error.ExceptionMapper;
 import com.bitmovin.analytics.features.Feature;
 import com.bitmovin.analytics.features.FeatureFactory;
+import com.bitmovin.analytics.license.FeatureConfigContainer;
 import com.bitmovin.analytics.stateMachines.PlayerState;
 import com.bitmovin.analytics.stateMachines.PlayerStateMachine;
 import com.bitmovin.analytics.utils.Util;
@@ -71,7 +72,7 @@ public class BitmovinSdkAdapter implements PlayerAdapter, EventDataManipulator {
         this.sourceMetadataMap = sourceMetadataMap;
     }
 
-    public Collection<Feature<?>> init() {
+    public Collection<Feature<FeatureConfigContainer, ?>> init() {
         resetSourceRelatedState();
         addPlayerListeners();
         checkAutoplayStartup();