+++ collector/src/main/java/com/bitmovin/analytics/ImpressionIdProvider.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics
+
+interface ImpressionIdProvider {
+    val impressionId: String
+}