+++ collector/src/main/java/com/bitmovin/analytics/features/FeatureFactory.kt
@@ -1,5 +1,7 @@
 package com.bitmovin.analytics.features
 
+import com.bitmovin.analytics.license.FeatureConfigContainer
+
 interface FeatureFactory {
-    fun createFeatures(): Collection<Feature<*>>
+    fun createFeatures(): Collection<Feature<FeatureConfigContainer, *>>
 }