+++ collector/src/main/java/com/bitmovin/analytics/features/segmenttracking/SegmentTracking.kt
@@ -1,21 +1,20 @@
 package com.bitmovin.analytics.features.segmenttracking
 
 import com.bitmovin.analytics.Observable
-import com.bitmovin.analytics.features.Feature
 import com.bitmovin.analytics.utils.QueueExtensions.Companion.limit
 import java.util.LinkedList
 import java.util.Queue
 
 class SegmentTracking(private vararg val observables: Observable<OnDownloadFinishedEventListener>) :
-        Feature<SegmentTrackingConfig>("segmentTracking", SegmentTrackingConfig::class),
         OnDownloadFinishedEventListener {
     companion object {
         const val defaultMaxSegments = 10
     }
     private val segmentQueue: Queue<Segment> = LinkedList()
 
-    val maxSegments
-        get() = config?.maxSegments ?: defaultMaxSegments
+    var maxSegments = defaultMaxSegments
+        private set
+
     val segments: Collection<Segment>
         get() = segmentQueue
 
@@ -23,12 +22,17 @@ class SegmentTracking(private vararg val observables: Observable<OnDownloadFinis
         observables.forEach { it.subscribe(this) }
     }
 
-    override fun configured(authenticated: Boolean, config: SegmentTrackingConfig?) {
+    fun configure(maxSegments: Int) {
+        this.maxSegments = maxSegments
         segmentQueue.limit(maxSegments)
     }
 
-    override fun disabled() {
+    fun disable() {
         observables.forEach { it.unsubscribe(this) }
+        segmentQueue.clear()
+    }
+
+    fun reset() {
         segmentQueue.clear()
     }