+++ collector/src/main/java/com/bitmovin/analytics/license/LicenseCallback.java
@@ -1,9 +1,7 @@
 package com.bitmovin.analytics.license;
 
-import java.util.Map;
-
 public interface LicenseCallback {
-    void configureFeatures(boolean authenticated, Map<String, String> settings);
+    void configureFeatures(boolean authenticated, FeatureConfigContainer featureConfigs);
 
     void authenticationCompleted(boolean success);
 }