+++ collector/src/test/java/com/bitmovin/analytics/features/errordetails/ErrorDetailTrackingTests.kt
@@ -16,15 +16,15 @@ import org.junit.Test
 class ErrorDetailTrackingTests {
     @Test
     fun testErrorDetailLimitSegmentsShouldntFailIfSegmentsAreNull() {
-        val errorDetail = ErrorDetail(0, null, null, null, null)
+        val errorDetail = ErrorDetail("", "", "", "", 0, 0, null, null, null, null)
         errorDetail.limitSegments(1)
     }
 
     @Test
     fun testErrorDetailLimitSegmentsShouldLimitSegments() {
-        val segment1 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, 0, true)
-        val segment2 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, 0, true)
-        val errorDetail = ErrorDetail(0, null, null, null, mutableListOf(segment1, segment2))
+        val segment1 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, null, 0, true)
+        val segment2 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, null, 0, true)
+        val errorDetail = ErrorDetail("", "", "", "", 0, 0, null, null, null, mutableListOf(segment1, segment2))
         assertThat(errorDetail.segments?.size).isEqualTo(2)
         errorDetail.limitSegments(1)
         assertThat(errorDetail.segments?.size).isEqualTo(1)
@@ -32,9 +32,9 @@ class ErrorDetailTrackingTests {
 
     @Test
     fun testErrorDetailLimitSegmentsShouldRemoveItemsFromEnd() {
-        val segment1 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, 0, true)
-        val segment2 = Segment(1, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, 0, true)
-        val errorDetail = ErrorDetail(0, null, null, null, mutableListOf(segment1, segment2))
+        val segment1 = Segment(0, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, null, 0, true)
+        val segment2 = Segment(1, SegmentType.MANIFEST_DASH, null, null, 0, 0.0, null, 0, true)
+        val errorDetail = ErrorDetail("", "", "", "", 0, 0, null, null, null, mutableListOf(segment1, segment2))
         errorDetail.limitSegments(1)
         assertThat(errorDetail.segments?.get(0)).isEqualTo(segment1)
         assertThat(errorDetail.segments?.get(0)).isNotEqualTo(segment2)
@@ -45,60 +45,49 @@ class ErrorDetailTrackingTests {
         val backend = mockk<ErrorDetailBackend>(relaxed = true)
         val support1 = ObservableSupport<OnErrorDetailEventListener>()
         val support2 = ObservableSupport<OnErrorDetailEventListener>()
-        val errorDetailTracking = ErrorDetailTracking(backend, null, support1, support2)
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, null, support1, support2)
         mockkObject(errorDetailTracking)
-        support1.notify { it.onError(0, null, null, null) }
-        support1.notify { it.onError(0, null, null, null) }
-        support2.notify { it.onError(0, null, null, null) }
-        support2.notify { it.onError(0, null, null, null) }
+        support1.notify { it.onError(null, null, null) }
+        support1.notify { it.onError(null, null, null) }
+        support2.notify { it.onError(null, null, null) }
+        support2.notify { it.onError(null, null, null) }
 
-        verify(exactly = 4) { errorDetailTracking.onError(any(), any(), any(), any()) }
+        verify(exactly = 4) { errorDetailTracking.onError(any(), any(), any()) }
     }
 
     @Test
     fun testSuccessfullyUnsubscribesFromSourcesAndClearsBackendOnDisabling() {
         val support = ObservableSupport<OnErrorDetailEventListener>()
         val backend = mockk<ErrorDetailBackend>(relaxed = true)
-        val errorDetailTracking = ErrorDetailTracking(backend, null, support)
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, null, support)
         mockkObject(errorDetailTracking)
-        support.notify { it.onError(0, null, null, null) }
-        verify { errorDetailTracking.onError(any(), any(), any(), any()) }
+        support.notify { it.onError(null, null, null) }
+        verify { errorDetailTracking.onError(any(), any(), any()) }
         clearMocks(errorDetailTracking)
         errorDetailTracking.disable()
         verify { backend.clear() }
-        support.notify { it.onError(0, null, null, null) }
-        verify(exactly = 0) { errorDetailTracking.onError(any(), any(), any(), any()) }
+        support.notify { it.onError(null, null, null) }
+        verify(exactly = 0) { errorDetailTracking.onError(any(), any(), any()) }
     }
 
     @Test
     fun testEnablesAndFlushesBackendAfterEnablingFeature() {
         val support = ObservableSupport<OnErrorDetailEventListener>()
         val backend = mockk<ErrorDetailBackend>(relaxed = true)
-        val errorDetailTracking = ErrorDetailTracking(backend, null, support)
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, null, support)
         errorDetailTracking.enabled()
         verify { backend.enabled = true }
         verify { backend.flush() }
     }
 
     @Test
-    fun testLimitsQueuedItemsAfterEnablingFeature() {
+    fun testLimitsQueuedItemsAfterConfiguringFeature() {
         val support = ObservableSupport<OnErrorDetailEventListener>()
         val backend = mockk<ErrorDetailBackend>(relaxed = true)
         val segmentTracking = SegmentTracking()
-        val errorDetailTracking = ErrorDetailTracking(backend, segmentTracking, support)
-        errorDetailTracking.enabled()
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, segmentTracking, support)
+        errorDetailTracking.configured(true, ErrorDetailTrackingConfig(true, 100))
         verify { backend.limitSegmentsInQueue(segmentTracking.maxSegments) }
-    }
-
-    @Test
-    fun testLimitsQueuedItemsToZeroIfSegmentTrackingIsDisabledAfterEnablingFeature() {
-        val support = ObservableSupport<OnErrorDetailEventListener>()
-        val backend = mockk<ErrorDetailBackend>(relaxed = true)
-        val segmentTracking = SegmentTracking()
-        segmentTracking.disable()
-        val errorDetailTracking = ErrorDetailTracking(backend, segmentTracking, support)
-        errorDetailTracking.enabled()
-        verify { backend.limitSegmentsInQueue(0) }
     }
 
     @Test
@@ -107,8 +96,8 @@ class ErrorDetailTrackingTests {
         val segmentTracking = SegmentTracking()
         segmentTracking.onDownloadFinished(OnDownloadFinishedEventObject(mockk()))
         segmentTracking.onDownloadFinished(OnDownloadFinishedEventObject(mockk()))
-        val errorDetailTracking = ErrorDetailTracking(backend, segmentTracking)
-        errorDetailTracking.onError(0, null, null, null)
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, segmentTracking)
+        errorDetailTracking.onError(null, null, null)
         val slot = slot<ErrorDetail>()
         verify { backend.send(capture(slot)) }
         assertThat(slot.captured.segments).isNotNull
@@ -122,8 +111,8 @@ class ErrorDetailTrackingTests {
         segmentTracking.onDownloadFinished(OnDownloadFinishedEventObject(mockk()))
         segmentTracking.onDownloadFinished(OnDownloadFinishedEventObject(mockk()))
         segmentTracking.disable()
-        val errorDetailTracking = ErrorDetailTracking(backend, segmentTracking)
-        errorDetailTracking.onError(0, null, null, null)
+        val errorDetailTracking = ErrorDetailTracking(mockk(relaxed = true), mockk(relaxed = true), mockk(relaxed = true), backend, segmentTracking)
+        errorDetailTracking.onError(null, null, null)
         val slot = slot<ErrorDetail>()
         verify { backend.send(capture(slot)) }
         assertThat(slot.captured.segments?.size ?: 0).isEqualTo(0)